Move schema to outside to avoid circular dependency
