+++ src/features/ingestion/adapters/schema.py
