+++ src/features/ingestion/parsing/__init__.py
@@ -2,7 +2,7 @@ from collections.abc import Iterator
 from pathlib import Path
 
 from src.features.ingestion import domain
-from src.features.ingestion.parsing import schema
+from src.features.ingestion.adapters import schema
 
 from . import _xml as xml
 from .academic_background import academic_background
