+++ src/features/ingestion/parsing/knowledge_areas.py
@@ -1,4 +1,4 @@
-from src.features.ingestion.parsing.schema import KnowledgeArea
+from src.features.ingestion.adapters.schema import KnowledgeArea
 
 from . import _xml as xml
